Actualizacion
union de las vistas y actualizacion de clases data

+++ src/entidades/Garante.java
@@ -11,22 +11,22 @@ public class Garante extends Persona {
         
     }
 
-    public Garante(TipoCliente tipo, String nombre, int dni, int cuilCuit, String domicilio, String ciudad, String codigoPostal, String lugarTrabajo, String telefono, String mail, boolean estado) {
+    public Garante(TipoCliente tipo, String nombre, int dni, long cuilCuit, String domicilio, String ciudad, String codigoPostal, String lugarTrabajo, String telefono, String mail, boolean estado) {
         super(tipo, nombre, dni, cuilCuit, domicilio, ciudad, codigoPostal, lugarTrabajo, telefono, mail, estado);
     }
 
-    public Garante(TipoCliente tipo, String nombre, int cuilCuit, String domicilio, String ciudad, String codigoPostal, String lugarTrabajo, String telefono, String mail, boolean estado) {
+    public Garante(TipoCliente tipo, String nombre, long cuilCuit, String domicilio, String ciudad, String codigoPostal, String lugarTrabajo, String telefono, String mail, boolean estado) {
         super(tipo, nombre, cuilCuit, domicilio, ciudad, codigoPostal, lugarTrabajo, telefono, mail, estado);
     }
     
 
-    public Garante(int idGarante, TipoCliente tipo, String nombre, int dni, int cuilCuit, String domicilio,
+    public Garante(int idGarante, TipoCliente tipo, String nombre, int dni, long cuilCuit, String domicilio,
             String ciudad, String codigoPostal, String lugarTrabajo, String telefono, String mail, boolean estado) {
         super(tipo, nombre, dni, cuilCuit, domicilio, ciudad, codigoPostal, lugarTrabajo, telefono, mail, estado);
         this.idGarante = idGarante;
     }
 
-    public Garante(int idGarante, TipoCliente tipo, String nombre, int cuilCuit, String domicilio, String ciudad,
+    public Garante(int idGarante, TipoCliente tipo, String nombre, long cuilCuit, String domicilio, String ciudad,
             String codigoPostal, String lugarTrabajo, String telefono, String mail, boolean estado) {
         super(tipo, nombre, cuilCuit, domicilio, ciudad, codigoPostal, lugarTrabajo, telefono, mail, estado);
         this.idGarante = idGarante;
